🚚 zb: move address to legacy_address
The following commits introduce a different implementation.
